moved hid patching code into input_proc, renamed injector to input_loader, removed transport_loop from input_proc and added similar function into injected.s
