--- injector/source/injected.s
@@ -1,27 +1,0 @@
-.global read_circlepad_and_ts
-.global read_circlepad_and_ts_sz
-
-read_circlepad_and_ts:
-STMFD   SP!, {R4-R6,LR}
-MOV     R5, R1
-MRC     p15, 0, R4,c13,c0, 3
-MOV     R1, #0x10000
-STR     R1, [R4,#0x80]!
-LDR     R0, [R0]
-SVC     0x32
-ANDS    R1, R0, #0x80000000
-BMI     .ret
-push {r3}
-ldr r3, =0x10df08
-LDRD    R0, [R4,#8]
-STRD    R0, [R3, #8]
-LDRD	r0, [R3]
-STRD	r0,	[R5]
-LDR     R0, [R4,#4]
-pop {r3}
-.ret:
-LDMFD   SP!, {R4-R6,PC}
-.LTORG # assembles literal pool
-
-read_circlepad_and_ts_sz:
-.4byte .-read_circlepad_and_ts
